Remove compatibility methods at BitBufferHelper
The BitBufferHelper methods getShort, getInt and getLong return a
short, int and long respectively. They have been replaced with
equivalent methods that return a Uint8, Uint16 or Uint32 types.
The following sections explain these replacements:

1. getShort
Code that required a Uint8 type instead of a short, started using
a new method "getUint8(byte[])". However a method that required to
collect the first N bits, where N is 16>N>8, started using a new
method "getUint16(byte[], int)".

2. getInt
Code that required a Uint16 type instead of an int, started using
a new method "getUint16(byte[])". However methods that required to
test data in 2 bytes, started using the "getUint16(byte[], int)"
method.

3. getLong
Code that required a Uint32 type instead of a long, started using
a new method "getUint32(byte[])".

+++ packethandler/implementation/src/main/java/org/opendaylight/l2switch/packethandler/decoders/ArpDecoder.java
@@ -66,8 +66,8 @@ public class ArpDecoder extends AbstractPacketDecoder<EthernetPacketReceived, Ar
                     KnownEtherType.forValue(BitBufferHelper.getInt(BitBufferHelper.getBits(data, bitOffset + 16, 16))));
 
             // Decode the hardware-length and protocol-length fields
-            builder.setHardwareLength(BitBufferHelper.getShort(BitBufferHelper.getBits(data, bitOffset + 32, 8)));
-            builder.setProtocolLength(BitBufferHelper.getShort(BitBufferHelper.getBits(data, bitOffset + 40, 8)));
+            builder.setHardwareLength(BitBufferHelper.getUint8(BitBufferHelper.getBits(data, bitOffset + 32, 8)));
+            builder.setProtocolLength(BitBufferHelper.getUint8(BitBufferHelper.getBits(data, bitOffset + 40, 8)));
 
             // Decode the operation field
             builder.setOperation(
